@@ -77,49 +77,49 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^12.2.9",
-    "@angular/cdk": "^12.2.9",
-    "@angular/common": "^12.2.9",
-    "@angular/compiler": "^12.2.9",
-    "@angular/core": "^12.2.9",
-    "@angular/forms": "^12.2.9",
-    "@angular/material": "^12.2.9",
-    "@angular/platform-browser": "^12.2.9",
-    "@angular/platform-browser-dynamic": "^12.2.9",
-    "@angular/router": "^12.2.9",
-    "@azure/msal-angular": "^2.0.1",
-    "@azure/msal-browser": "^2.15.0",
+    "@angular/animations": "^14.1.2",
+    "@angular/cdk": "^14.1.1",
+    "@angular/common": "^14.1.2",
+    "@angular/compiler": "^14.1.2",
+    "@angular/core": "^14.1.2",
+    "@angular/forms": "^14.1.2",
+    "@angular/material": "^14.1.1",
+    "@angular/platform-browser": "^14.1.2",
+    "@angular/platform-browser-dynamic": "^14.1.2",
+    "@angular/router": "^14.1.2",
+    "@azure/msal-angular": "^2.4.1",
+    "@azure/msal-browser": "^2.28.1",
     "@ionic-native/core": "^5.10.0",
     "@ionic-native/file": "^5.6.0",
     "@ionic-native/file-opener": "^5.6.0",
     "@ionic-native/network": "^5.12.0",
-    "@ionic/angular": "^5.2.3",
-    "@mat-datetimepicker/core": "~7.0.1",
+    "@ionic/angular": "^6.1.14",
+    "@mat-datetimepicker/core": "~10.0.0",
     "@mdi/angular-material": "^6.2.95",
-    "@ngx-translate/core": "^13.0.0",
+    "@ngx-translate/core": "^14.0.0",
     "@turf/buffer": "^6.5.0",
     "@turf/helpers": "^6.5.0",
     "@turf/line-intersect": "^6.5.0",
     "@turf/point-on-feature": "^6.5.0",
-    "angular-shepherd": "12.0.0",
+    "ajv": "^8.11.0",
+    "angular-shepherd": "14.0.0",
     "bowser": "^2.10.0",
-    "classlist.js": "^1.1.20150312",
-    "core-js": "^3.6.5",
+    "core-js": "^3.24.1",
     "file-saver": "^2.0.2",
     "hammerjs": "^2.0.8",
     "html2canvas": "^1.0.0-rc.5",
-    "jspdf": "^2.4.0",
-    "jszip": "^3.1.5",
+    "jspdf": "^2.5.1",
+    "jszip": "^3.10.1",
     "jwt-decode": "^2.2.0",
     "moment": "^2.29.4",
     "ngx-color-picker": "^10.1.0",
     "ngx-indexed-db": "^11.0.2",
-    "ngx-toastr": "^14.1.3",
+    "ngx-toastr": "^15.0.0",
     "nosleep.js": "^0.12.0",
-    "ol": "^6.8.1",
-    "ol-mapbox-style": "^6.5.1",
-    "proj4": "^2.7.5",
-    "rxjs": "^6.6.7",
+    "ol": "^6.15.1",
+    "ol-mapbox-style": "^8.2.1",
+    "proj4": "^2.8.0",
+    "rxjs": "^7.5.6",
     "scroll-into-view-if-needed": "^2.2.20",
     "scss-bundle": "^3.1.2",
     "stream": "^0.0.2",
@@ -129,38 +129,38 @@
     "tslib": "^2.0.0",
     "typy": "^3.3.0",
     "unorm": "^1.4.1",
-    "web-animations-js": "^2.3.2",
     "windows-1252": "^3.0.4",
-    "zone.js": "~0.11.4"
+    "zone.js": "~0.11.7"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^12.2.9",
-    "@angular-eslint/builder": "12.5.0",
-    "@angular-eslint/eslint-plugin": "12.5.0",
-    "@angular-eslint/eslint-plugin-template": "12.5.0",
-    "@angular-eslint/schematics": "12.5.0",
-    "@angular-eslint/template-parser": "12.5.0",
-    "@angular/cli": "^12.2.9",
-    "@angular/compiler-cli": "^12.2.9",
-    "@babel/core": "^7.7.4",
-    "@babel/preset-env": "^7.7.4",
+    "@angular-devkit/build-angular": "^14.1.2",
+    "@angular-eslint/builder": "14.0.2",
+    "@angular-eslint/eslint-plugin": "14.0.2",
+    "@angular-eslint/eslint-plugin-template": "14.0.2",
+    "@angular-eslint/schematics": "14.0.2",
+    "@angular-eslint/template-parser": "14.0.2",
+    "@angular/cli": "^14.1.2",
+    "@angular/compiler-cli": "^14.1.2",
+    "@babel/core": "^7.18.10",
+    "@babel/preset-env": "^7.18.10",
     "@compodoc/compodoc": "^1.1.15",
-    "@ngx-translate/http-loader": "^6.0.0",
-    "@types/geojson": "^7946.0.6",
+    "@ngx-translate/http-loader": "^7.0.0",
+    "@octokit/core": "^4.0.0",
+    "@types/geojson": "^7946.0.10",
     "@types/hammerjs": "^2.0.36",
     "@types/html2canvas": "0.0.35",
     "@types/jasmine": "^3.5.11",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^14.0.22",
-    "@typescript-eslint/eslint-plugin": "^4.28.2",
-    "@typescript-eslint/parser": "^4.28.2",
-    "angular-cli-ghpages": "^1.0.0-rc.2",
+    "@types/node": "^14.18.23",
+    "@typescript-eslint/eslint-plugin": "^5.33.0",
+    "@typescript-eslint/parser": "^5.33.0",
+    "angular-cli-ghpages": "^1.0.3",
     "conventional-changelog-cli": "^2.0.12",
     "copy-newer": "^2.1.2",
     "del": "^6.0.0",
-    "eslint": "^7.26.0",
-    "eslint-plugin-unused-imports": "^1.1.5",
-    "gh-release": "^6.0.1",
+    "eslint": "^8.21.0",
+    "eslint-plugin-unused-imports": "^2.0.0",
+    "gh-release": "^6.0.4",
     "gulp": "^4.0.0",
     "gulp-babel": "^8.0.0",
     "gulp-concat": "^2.6.1",
@@ -172,19 +172,19 @@
     "gulp-wait": "0.0.2",
     "jasmine-core": "~3.8.0",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "^6.3.16",
+    "karma": "^6.4.0",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~3.0.3",
+    "karma-coverage": "^2.2.0",
     "karma-jasmine": "~4.0.1",
     "karma-jasmine-html-reporter": "^1.7.0",
-    "ng-packagr": "^12.2.3",
+    "ng-packagr": "^14.1.0",
     "npm-run-all": "^4.1.5",
     "protractor": "~7.0.0",
-    "sass": "^1.42.1",
-    "ts-node": "^10.1.0",
-    "typescript": "~4.3.5",
+    "sass": "^1.54.4",
+    "ts-node": "^10.9.1",
+    "typescript": "~4.6.4",
     "wait-on": "^6.0.0",
     "webdriver-manager": "^12.1.6",
-    "webpack": "^5.58.1"
+    "webpack": "^5.74.0"
   }
 }
