@@ -1,6 +1,6 @@
 {
   "name": "@igo2/auth",
-  "version": "1.12.1",
+  "version": "1.13.0-alpha.0",
   "description": "IGO Library",
   "author": "MSP",
   "keywords": ["igo"],
@@ -20,13 +20,13 @@
     "tslib": "^2.0.0"
   },
   "peerDependencies": {
-    "@angular/cdk": "^12.2.9",
-    "@angular/common": "^12.2.9",
-    "@angular/core": "^12.2.9",
-    "@angular/forms": "^12.2.9",
-    "@angular/material": "^12.2.9",
-    "@igo2/core": "^1.12.1",
-    "@igo2/utils": "^1.12.1",
+    "@angular/cdk": "^14.0.5",
+    "@angular/common": "^14.0.6",
+    "@angular/core": "^14.0.6",
+    "@angular/forms": "^14.0.6",
+    "@angular/material": "^14.0.5",
+    "@igo2/core": "^1.13.0-alpha.0",
+    "@igo2/utils": "^1.13.0-alpha.0",
     "rxjs": "^6.6.7",
     "@azure/msal-angular": "^2.0.0",
     "@azure/msal-browser": "^2.14.2"
