@@ -1,6 +1,6 @@
 {
   "name": "@igo2/context",
-  "version": "1.12.1",
+  "version": "1.13.0-alpha.0",
   "description": "IGO Library",
   "author": "MSP",
   "keywords": ["igo"],
@@ -14,17 +14,17 @@
     "url": "https://github.com/infra-geo-ouverte/igo2-lib/issues"
   },
   "peerDependencies": {
-    "@angular/animations": "^12.2.9",
-    "@angular/common": "^12.2.9",
-    "@angular/core": "^12.2.9",
-    "@angular/forms": "^12.2.9",
-    "@angular/material": "^12.2.9",
-    "@igo2/auth": "^1.12.1",
-    "@igo2/common": "^1.12.1",
-    "@igo2/core": "^1.12.1",
-    "@igo2/utils": "^1.12.1",
-    "@igo2/geo": "^1.12.1",
-    "ol": "^6.8.1",
+    "@angular/animations": "^14.0.6",
+    "@angular/common": "^14.0.6",
+    "@angular/core": "^14.0.6",
+    "@angular/forms": "^14.0.6",
+    "@angular/material": "^14.0.5",
+    "@igo2/auth": "^1.13.0-alpha.0",
+    "@igo2/common": "^1.13.0-alpha.0",
+    "@igo2/core": "^1.13.0-alpha.0",
+    "@igo2/utils": "^1.13.0-alpha.0",
+    "@igo2/geo": "^1.13.0-alpha.0",
+    "ol": "~6.14.1",
     "rxjs": "^6.6.7"
   },
   "dependencies": {
