@@ -1,6 +1,6 @@
 {
   "name": "@igo2/geo",
-  "version": "1.12.1",
+  "version": "1.13.0-alpha.0",
   "description": "IGO Library",
   "author": "MSP",
   "keywords": ["igo"],
@@ -24,24 +24,24 @@
     "windows-1252": "^3.0.4"
   },
   "peerDependencies": {
-    "@angular/animations": "^12.2.9",
-    "@angular/common": "^12.2.9",
-    "@angular/core": "^12.2.9",
-    "@angular/forms": "^12.2.9",
-    "@angular/material": "^12.2.9",
-    "@igo2/common": "^1.12.1",
-    "@igo2/core": "^1.12.1",
-    "@igo2/utils": "^1.12.1",
-    "ol": "^6.8.1",
-    "proj4": "^2.7.5",
+    "@angular/animations": "^14.0.6",
+    "@angular/common": "^14.0.6",
+    "@angular/core": "^14.0.6",
+    "@angular/forms": "^14.0.6",
+    "@angular/material": "^14.0.5",
+    "@igo2/common": "^1.13.0-alpha.0",
+    "@igo2/core": "^1.13.0-alpha.0",
+    "@igo2/utils": "^1.13.0-alpha.0",
+    "ol": "~6.14.1",
+    "proj4": "^2.8.0",
     "rxjs": "^6.6.7",
-    "ol-mapbox-style": "^6.4.1",
+    "ol-mapbox-style": "^8.2.0",
     "striptags": "^3.2.0"
   },
   "optionalDependencies": {
     "file-saver": "^2.0.2",
     "html2canvas": "^1.0.0-rc.5",
-    "jspdf": "^2.3.1",
+    "jspdf": "^2.5.1",
     "jszip": "^3.1.5",
     "moment": "^2.24.0",
     "stream": "^0.0.2"
