@@ -1,6 +1,6 @@
 {
   "name": "@igo2/utils",
-  "version": "1.12.1",
+  "version": "1.13.0-alpha.0",
   "description": "IGO Library",
   "author": "MSP",
   "keywords": ["igo"],
@@ -18,7 +18,7 @@
     "tslib": "^2.0.0"
   },
   "peerDependencies": {
-    "@angular/common": "^12.2.9",
-    "@angular/core": "^12.2.9"
+    "@angular/common": "^14.0.6",
+    "@angular/core": "^14.0.6"
   }
 }
