@@ -1,4 +1,8 @@
-import { Component, Input } from '@angular/core';
+import { Component, Input, ChangeDetectionStrategy, ChangeDetectorRef, OnDestroy, OnInit } from '@angular/core';
+import { MatIconRegistry } from '@angular/material/icon';
+import { Observable, forkJoin, Subject } from 'rxjs';
+import { tap, take, takeUntil } from 'rxjs/operators';
+
 import {
   IgoMap,
   DataSourceService,
@@ -10,28 +14,35 @@ import {
   featureToOl,
   DataSource,
   QueryableDataSourceOptions,
-  Layer,
   SpatialFilterService,
-  SpatialFilterQueryType,
   SpatialFilterType,
   SpatialFilterItemType,
-  SpatialFilterThematic
+  SpatialFilterQueryType,
+  SpatialFilterThematic,
+  Layer,
+  VectorLayer,
+  createOverlayMarkerStyle,
+  MeasureLengthUnit
 } from '@igo2/geo';
-import { MessageService, LanguageService } from '@igo2/core';
 import { EntityStore } from '@igo2/common';
 import olFormatGeoJSON from 'ol/format/GeoJSON';
+import olSourceVector from 'ol/source/Vector';
+import olSourceCluster from 'ol/source/Cluster';
+import type { default as OlGeometry } from 'ol/geom/Geometry';
 import { BehaviorSubject } from 'rxjs';
 import * as olstyle from 'ol/style';
-import olSourceCluster from 'ol/source/Cluster';
-import olSourceVector from 'ol/source/Vector';
-import type { default as OlGeometry } from 'ol/geom/Geometry';
+import { MessageService, LanguageService } from '@igo2/core';
 
+/**
+ * Spatial Filter Type
+ */
 @Component({
   selector: 'app-spatial-filter',
   templateUrl: './spatial-filter.component.html',
-  styleUrls: ['./spatial-filter.component.scss']
+  styleUrls: ['./spatial-filter.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class AppSpatialFilterComponent {
+export class AppSpatialFilterComponent implements OnInit, OnDestroy {
   public map = new IgoMap({
     controls: {
       attribution: {
@@ -40,9 +51,9 @@ export class AppSpatialFilterComponent {
     }
   });
 
-  view = {
-    center: [-71.9, 46.9],
-    zoom: 10
+  public view = {
+    center: [-73, 47.2],
+    zoom: 6
   };
 
   @Input() type: SpatialFilterType;
@@ -50,11 +61,15 @@ export class AppSpatialFilterComponent {
   @Input() freehandDrawIsActive: boolean;
 
   public layers: Layer[] = [];
+  public activeLayers: Layer[] = [];
 
   public queryType: SpatialFilterQueryType;
   public thematics: SpatialFilterThematic[];
   public zone: Feature;
-  public radius: number;
+  public zoneWithBuffer: Feature;
+  public buffer = 0;
+
+  public iterator = 1;
 
   public selectedFeature$: BehaviorSubject<Feature> = new BehaviorSubject(
     undefined
@@ -62,18 +77,27 @@ export class AppSpatialFilterComponent {
 
   private format = new olFormatGeoJSON();
 
-  public store = new EntityStore<Feature>([]); // Store to print results at the end
+  public store: EntityStore<Feature> = new EntityStore<Feature>([]); // Store to print results at the end
 
-  public spatialListStore = new EntityStore<Feature>([]);
+  public spatialListStore: EntityStore<Feature> = new EntityStore<Feature>([]);
 
   public loading = false;
 
+  public thematicLength = 0;
+
+  public measureUnit: MeasureLengthUnit = MeasureLengthUnit.Meters;
+  private unsubscribe$ = new Subject<void>();
+
+  public defaultStyle: olstyle.Style | ((feature, resolution) => olstyle.Style);
+
   constructor(
+    private matIconRegistry: MatIconRegistry,
     private spatialFilterService: SpatialFilterService,
     private dataSourceService: DataSourceService,
     private layerService: LayerService,
     private messageService: MessageService,
-    private languageService: LanguageService
+    private languageService: LanguageService,
+    private cdRef: ChangeDetectorRef
   ) {
     this.dataSourceService
       .createAsyncDataSource({
@@ -83,18 +107,28 @@ export class AppSpatialFilterComponent {
         this.map.addLayer(
           this.layerService.createLayer({
             title: 'OSM',
-            source: dataSource,
-            baseLayer: true,
-            visible: true
+            source: dataSource
           })
         );
       });
   }
 
+  ngOnInit() {
+    for (const layer of this.map.layers) {
+      if (layer.title && layer.title.includes(this.languageService.translate.instant('igo.geo.spatialFilter.spatialFilter'))) {
+        this.layers.push(layer);
+      }
+    }
+  }
+
+  ngOnDestroy() {
+    this.unsubscribe$.next();
+    this.unsubscribe$.complete();
+  }
+
   getOutputType(event: SpatialFilterType) {
     this.type = event;
     this.queryType = undefined;
-    this.radius = undefined;
   }
 
   getOutputQueryType(event: SpatialFilterQueryType) {
@@ -107,6 +141,7 @@ export class AppSpatialFilterComponent {
   private loadFilterList() {
     this.spatialFilterService
       .loadFilterList(this.queryType)
+      .pipe(takeUntil(this.unsubscribe$))
       .subscribe((features: Feature[]) => {
         features.sort((a, b) => {
           if (a.properties.nom < b.properties.nom) {
@@ -131,75 +166,117 @@ export class AppSpatialFilterComponent {
     this.queryType = undefined;
   }
 
+  clearMap() {
+    this.map.removeLayers(this.layers);
+    this.layers = [];
+    this.activeLayers = [];
+    this.thematicLength = 0;
+    this.iterator = 1;
+    if (this.type === SpatialFilterType.Predefined) {
+      this.zone = undefined;
+      this.queryType = undefined;
+    }
+  }
+
   private loadThematics() {
     this.loading = true;
-    this.tryAddFeaturesToMap([this.zone]);
-    if (!this.thematics) {
+    let zeroResults = true;
+    let thematics;
+    if (this.buffer === 0 || this.type === SpatialFilterType.Point) {
+      this.tryAddFeaturesToMap([this.zone]);
+    }
+    if (this.itemType !== SpatialFilterItemType.Thematics) {
       const theme: SpatialFilterThematic = {
         name: ''
       };
-      this.thematics = [theme];
+      thematics = [theme];
+    } else {
+      thematics = this.thematics;
+    }
+    if (this.measureUnit === MeasureLengthUnit.Kilometers && this.type !== SpatialFilterType.Point) {
+      this.buffer = this.buffer * 1000;
     }
     if (this.type === SpatialFilterType.Polygon) {
-      this.radius = undefined;
+      this.buffer = 0; // to avoid buffer enter a second time in terrAPI
     }
-    this.thematics.forEach(thematic => {
-      this.spatialFilterService
-        .loadFilterItem(
-          this.zone,
-          this.itemType,
-          this.queryType,
-          thematic,
-          this.radius
-        )
-        .subscribe((features: Feature[]) => {
-          this.store.insertMany(features);
-          const featuresPoint: Feature[] = [];
-          const featuresLinePoly: Feature[] = [];
-          let idPoint;
-          let idLinePoly;
-          features.forEach(feature => {
-            if (feature.geometry.type === 'Point') {
-              featuresPoint.push(feature);
-              idPoint = feature.meta.id;
-            } else {
-              featuresLinePoly.push(feature);
-              idLinePoly = feature.meta.id;
-            }
-          });
-          this.tryAddPointToMap(featuresPoint, idPoint);
-          this.tryAddLayerToMap(featuresLinePoly, idLinePoly);
-          this.loading = false;
-          if (features.length >= 10000) {
-            this.messageService.alert(
-              this.languageService.translate.instant(
-                'igo.geo.spatialFilter.maxSizeAlert'
-              ),
-              this.languageService.translate.instant(
-                'igo.geo.spatialFilter.warning'
-              ),
-              { timeOut: 10000 }
-            );
-          }
-          if (!features.length) {
-            this.messageService.alert(
-              this.languageService.translate.instant(
-                'igo.geo.spatialFilter.zeroResults'
-              ),
-              this.languageService.translate.instant(
-                'igo.geo.spatialFilter.warning'
-              ),
-              { timeOut: 10000 }
-            );
-          }
-        });
+
+    const observables$: Observable<Feature[]>[] = [];
+    thematics.forEach(thematic => {
+      observables$.push(
+        this.spatialFilterService
+          .loadFilterItem(
+            this.zone,
+            this.itemType,
+            this.queryType,
+            thematic,
+            this.buffer
+          )
+          .pipe(
+            tap((features: Feature[]) => {
+              this.store.insertMany(features);
+              const featuresPoint: Feature[] = [];
+              const featuresLinePoly: Feature[] = [];
+              let idPoint;
+              let idLinePoly;
+              features.forEach(feature => {
+                if (feature.geometry.type === 'Point') {
+                  feature.properties.longitude = feature.geometry.coordinates[0];
+                  feature.properties.latitude = feature.geometry.coordinates[1];
+                  featuresPoint.push(feature);
+                  idPoint = feature.meta.id;
+                } else {
+                  featuresLinePoly.push(feature);
+                  idLinePoly = feature.meta.id;
+                }
+              });
+
+              this.tryAddPointToMap(featuresPoint, idPoint);
+              this.tryAddLayerToMap(featuresLinePoly, idLinePoly);
+              if (features.length) {
+                zeroResults = false;
+                this.thematicLength += 1;
+                thematic.zeroResults = false;
+                this.cdRef.detectChanges();
+              } else {
+                thematic.zeroResults = true;
+              }
+
+              if (features.length >= 10000) {
+                this.messageService.alert(
+                  this.languageService.translate.instant(
+                    'igo.geo.spatialFilter.maxSizeAlert'
+                  ),
+                  this.languageService.translate.instant(
+                    'igo.geo.spatialFilter.warning'
+                  ),
+                  { timeOut: 10000 }
+                );
+              }
+            })
+          )
+      );
+    });
+
+    forkJoin(observables$).pipe(takeUntil(this.unsubscribe$)).subscribe(() => {
+      this.loading = false;
+      if (zeroResults) {
+        this.messageService.alert(
+          this.languageService.translate.instant(
+            'igo.geo.spatialFilter.zeroResults'
+          ),
+          this.languageService.translate.instant(
+            'igo.geo.spatialFilter.warning'
+          ),
+          { timeOut: 10000 }
+        );
+      }
     });
   }
 
-  onZoneChange(feature: Feature) {
+  onZoneChange(feature: Feature, buffer?: boolean) {
     this.zone = feature;
     if (feature) {
-      this.tryAddFeaturesToMap([feature]);
+      buffer ? this.tryAddFeaturesToMap([feature], true) : this.tryAddFeaturesToMap([feature]);
       this.zoomToFeatureExtent(feature);
     }
   }
@@ -207,35 +284,85 @@ export class AppSpatialFilterComponent {
   /**
    * Try to add zone feature to the map overlay
    */
-  public tryAddFeaturesToMap(features: Feature[]) {
+  public tryAddFeaturesToMap(features: Feature[], buffer?: boolean) {
     let i = 1;
     for (const feature of features) {
       if (this.type === SpatialFilterType.Predefined) {
-        for (const layer of this.map.layers) {
+        for (const layer of this.layers) {
           if (
             layer.options._internal &&
-            layer.options._internal.code === feature.properties.code
+            layer.options._internal.code === feature.properties.code &&
+            !buffer
           ) {
+            if (!layer.title?.startsWith('Zone')) {
+              const index = this.layers.indexOf(layer);
+              this.layers.splice(index, 1);
+            }
             return;
           }
-          if (layer.title.startsWith('Zone')) {
+          if (layer.title?.startsWith('Zone')) {
+            this.activeLayers = [];
+            const index = this.layers.indexOf(layer);
+            this.layers.splice(index, 1);
             this.map.removeLayer(layer);
           }
         }
+      } else {
+        if (buffer) {
+          for (const layer of this.activeLayers) {
+            if (this.activeLayers.length === 1 && layer.title?.startsWith('Zone')) {
+              const index = this.layers.indexOf(layer);
+              this.layers.splice(index, 1);
+              this.map.removeLayer(layer);
+            }
+          }
+        }
+        this.activeLayers = [];
       }
-      for (const layer of this.map.layers) {
-        if (layer.title.startsWith('Zone')) {
+      for (const layer of this.layers) {
+        if (layer.title?.startsWith('Zone')) {
           i++;
         }
       }
+      this.defaultStyle = (_feature, resolution) => {
+        const coordinates = (features[0] as any).coordinates;
+        return new olstyle.Style({
+          image: new olstyle.Circle({
+            radius: coordinates
+              ? this.buffer /
+                Math.cos((Math.PI / 180) * coordinates[1]) /
+                resolution
+              : undefined,
+            fill: new olstyle.Fill({
+              color: 'rgba(200, 200, 20, 0.2)'
+            }),
+            stroke: new olstyle.Stroke({
+              width: 1,
+              color: 'orange'
+            })
+          }),
+          stroke: new olstyle.Stroke({
+            width: 1,
+            color: 'orange'
+          }),
+          fill: new olstyle.Fill({
+            color: 'rgba(200, 200, 20, 0.2)'
+          })
+        });
+      };
       this.dataSourceService
         .createAsyncDataSource({
           type: 'vector',
           queryable: true
         } as QueryableDataSourceOptions)
+        .pipe(take(1))
         .subscribe((dataSource: DataSource) => {
           const olLayer = this.layerService.createLayer({
-            title: ('Zone ' + i) as string,
+            isIgoInternalLayer: true,
+            title: ('Zone ' + i + ' - ' + this.languageService.translate.instant(
+              'igo.geo.spatialFilter.spatialFilter'
+            )) as string,
+            workspace: { enabled: true },
             _internal: {
               code:
                 this.type === SpatialFilterType.Predefined
@@ -243,57 +370,39 @@ export class AppSpatialFilterComponent {
                   : undefined
             },
             source: dataSource,
-            visible: true,
-            style: (_feature, resolution) => {
-              const coordinates = (features[0] as any).coordinates;
-              return new olstyle.Style({
-                image: new olstyle.Circle({
-                  radius: coordinates
-                    ? this.radius /
-                      Math.cos((Math.PI / 180) * coordinates[1]) /
-                      resolution
-                    : undefined,
-                  fill: new olstyle.Fill({
-                    color: 'rgba(200, 200, 20, 0.2)'
-                  }),
-                  stroke: new olstyle.Stroke({
-                    width: 1,
-                    color: 'orange'
-                  })
-                }),
-                stroke: new olstyle.Stroke({
-                  width: 1,
-                  color: 'orange'
-                }),
-                fill: new olstyle.Fill({
-                  color: 'rgba(200, 200, 20, 0.2)'
-                })
-              });
-            }
-          });
+            visible: true
+          }) as VectorLayer;
           const featuresOl = features.map(f => {
             return featureToOl(f, this.map.projection);
           });
+          if (this.type !== SpatialFilterType.Predefined) {
+            const type = this.type === SpatialFilterType.Point ? 'Cercle' : 'Polygone';
+            featuresOl[0].set('nom', 'Zone', true);
+            featuresOl[0].set('type', type, true);
+          }
           const ol = dataSource.ol as olSourceVector<OlGeometry> | olSourceCluster;
           ol.addFeatures(featuresOl);
+          olLayer.ol.setStyle(this.defaultStyle);
           this.map.addLayer(olLayer);
           this.layers.push(olLayer);
+          this.activeLayers.push(olLayer);
+          this.cdRef.detectChanges();
         });
     }
   }
 
   /**
-   * Try to point features to the map
+   * Try to add point features to the map
    * Necessary to create clusters
    */
   private tryAddPointToMap(features: Feature[], id) {
     let i = 1;
-    if (features.length >= 1) {
+    if (features.length) {
       if (this.map === undefined) {
         return;
       }
-      for (const layer of this.map.layers) {
-        if (layer.title.startsWith(features[0].meta.title)) {
+      for (const layer of this.layers) {
+        if (layer.title?.startsWith(features[0].meta.title)) {
           i++;
         }
       }
@@ -307,45 +416,78 @@ export class AppSpatialFilterComponent {
             title: 'Cluster'
           }
         } as QueryableDataSourceOptions)
+        .pipe(take(1))
         .subscribe((dataSource: ClusterDataSource) => {
+          const icon = features[0].meta.icon;
+          let style: olstyle.Style;
+          if (!icon) {
+            style = createOverlayMarkerStyle();
+          } else {
+            style = this.createSvgIcon(icon) || createOverlayMarkerStyle();
+          }
+
           const olLayer = this.layerService.createLayer({
-            title: (features[0].meta.title + ' ' + i) as string,
+            isIgoInternalLayer: true,
+            title: (features[0].meta.title + ' ' + i + ' - ' + this.languageService.translate.instant(
+              'igo.geo.spatialFilter.spatialFilter'
+            )) as string,
             source: dataSource,
             visible: true,
-            clusterParam: {
-              clusterRanges: [{ minRadius: 1, maxRadius: 5, style: {} }]
-            }
+            style
           });
+
           const featuresOl = features.map(feature => {
             return featureToOl(feature, this.map.projection);
           });
           const ol = dataSource.ol as olSourceCluster;
           ol.getSource().addFeatures(featuresOl);
-          if (this.map.layers.find(layer => layer.id === olLayer.id)) {
+          if (this.layers.find(layer => layer.id === olLayer.id)) {
             this.map.removeLayer(
-              this.map.layers.find(layer => layer.id === olLayer.id)
+              this.layers.find(layer => layer.id === olLayer.id)
             );
             i = i - 1;
-            olLayer.title = (features[0].meta.title + ' ' + i) as string;
+            olLayer.title = (features[0].meta.title + ' ' + i + ' - ' + this.languageService.translate.instant(
+              'igo.geo.spatialFilter.spatialFilter'
+            )) as string;
             olLayer.options.title = olLayer.title;
           }
+          this.iterator = i;
           this.map.addLayer(olLayer);
           this.layers.push(olLayer);
+          this.pushLayer(olLayer);
+          this.cdRef.detectChanges();
         });
     }
   }
 
+  private createSvgIcon(icon): olstyle.Style {
+    let style: olstyle.Style;
+    this.matIconRegistry.getNamedSvgIcon(icon).subscribe(svgObj => {
+      const xmlSerializer = new XMLSerializer();
+      svgObj.setAttribute('width', '30');
+      svgObj.setAttribute('height', '30');
+      svgObj.setAttribute('fill', 'rgba(0, 128, 255)');
+      svgObj.setAttribute('stroke', 'white');
+      const svg = xmlSerializer.serializeToString(svgObj);
+      style = new olstyle.Style({
+        image: new olstyle.Icon({
+          src: 'data:image/svg+xml;utf8,' + svg
+        })
+      });
+    });
+    return style;
+  }
   /**
    * Try to add line or polygon features to the map
    */
   private tryAddLayerToMap(features: Feature[], id) {
     let i = 1;
-    if (features.length > 1) {
+    if (features.length) {
       if (this.map === undefined) {
         return;
       }
-      for (const layer of this.map.layers) {
-        if (layer.title.startsWith(features[0].meta.title)) {
+      for (const layer of this.layers) {
+        if (layer.title?.startsWith(features[0].meta.title)) {
           i++;
         }
       }
@@ -355,27 +497,35 @@ export class AppSpatialFilterComponent {
           id,
           queryable: true
         } as QueryableDataSourceOptions)
+        .pipe(take(1))
         .subscribe((dataSource: DataSource) => {
           const olLayer = this.layerService.createLayer({
-            title: (features[0].meta.title + ' ' + i) as string,
+            isIgoInternalLayer: true,
+            title: (features[0].meta.title + ' ' + i + ' - ' + this.languageService.translate.instant(
+              'igo.geo.spatialFilter.spatialFilter'
+            )) as string,
             source: dataSource,
             visible: true
           });
           const featuresOl = features.map(feature => {
             return featureToOl(feature, this.map.projection);
           });
-          const ol = dataSource.ol as olSourceVector<OlGeometry> | olSourceCluster;
+          const ol = dataSource.ol as olSourceVector<OlGeometry>;
           ol.addFeatures(featuresOl);
-          if (this.map.layers.find(layer => layer.id === olLayer.id)) {
+          if (this.layers.find(layer => layer.id === olLayer.id)) {
             this.map.removeLayer(
-              this.map.layers.find(layer => layer.id === olLayer.id)
+              this.layers.find(layer => layer.id === olLayer.id)
             );
             i = i - 1;
-            olLayer.title = (features[0].meta.title + ' ' + i) as string;
+            olLayer.title = (features[0].meta.title + ' ' + i + ' - ' + this.languageService.translate.instant(
+              'igo.geo.spatialFilter.spatialFilter'
+            )) as string;
             olLayer.options.title = olLayer.title;
           }
           this.map.addLayer(olLayer);
           this.layers.push(olLayer);
+          this.pushLayer(olLayer);
+          this.cdRef.detectChanges();
         });
     }
   }
@@ -390,15 +540,13 @@ export class AppSpatialFilterComponent {
     }
   }
 
-  /**
-   * Permit the query action on one item result
-   */
-  handleQueryResults(results) {
-    const features: Feature[] = results.features;
-    let feature;
-    if (features.length) {
-      feature = features[0];
+  pushLayer(layer) {
+    for (const lay of this.activeLayers) {
+      if (lay.id === layer.id) {
+        return;
+      }
     }
-    this.selectedFeature$.next(feature);
+
+    this.activeLayers.push(layer);
   }
 }
