@@ -18,25 +18,14 @@
 
 /*******************************************************************************
  * BROWSER POLYFILLS
- */
-
-/** IE10 and IE11 requires the following for NgClass support on SVG elements */
-import 'classlist.js'; // Run `npm install --save classlist.js`.
+ */ // Run `npm install --save classlist.js`.
 
 /** IE10 and IE11 requires the following for string normalization */
 import 'unorm/lib/unorm.js';
 
 /** IE10 and IE11 requires the following for some object management */
 import 'core-js/es/object';
-import 'core-js/es/string';
-/**
- * Web Animations `@angular/platform-browser/animations`
- * Only required if AnimationBuilder is used within the application and
- * using IE/Edge or Safari.
- * Standard animation support in Angular DOES NOT require any polyfills (as of
- * Angular 6.0).
- */
-import 'web-animations-js'; // Run `npm install --save web-animations-js`.
+import 'core-js/es/string'; // Run `npm install --save web-animations-js`.
 
 /**
  * By default, zone.js will patch all possible macroTask and DomEvents
