@@ -1,6 +1,7 @@
 import olSource from 'ol/source/Source';
 import olSourceVector from 'ol/source/Vector';
 import type { default as OlGeometry } from 'ol/geom/Geometry';
+import type { ServerType } from 'ol/source/wms';
 
 import { DataSourceOptions } from './datasource.interface';
 import { WFSDataSourceOptionsParams } from './wfs-datasource.interface';
@@ -14,7 +15,7 @@ export interface WMSDataSourceOptions extends DataSourceOptions {
   crossOrigin?: string;
   projection?: string;
   resolutions?: number[];
-  serverType?: string;
+  serverType?: ServerType;
   ratio?: number;
   ol?: olSourceVector<OlGeometry> | olSource;
   refreshIntervalSec?: number;
