@@ -1,4 +1,4 @@
-import type { default as OlGeometryType } from 'ol/geom/GeometryType';
+import type { Type } from 'ol/geom/Geometry';
 import { GeoJsonGeometryTypes } from 'geojson';
 
 import { FormFieldInputs } from '@igo2/common';
@@ -7,7 +7,7 @@ import { IgoMap } from '../../map';
 
 export interface GeometryFormFieldInputs extends FormFieldInputs {
   map: IgoMap;
-  geometryType: typeof OlGeometryType;
+  geometryType: Type;
   tooltip?: string;
 }
 
