@@ -5,7 +5,6 @@ import { unByKey } from 'ol/Observable';
 import { easeOut } from 'ol/easing';
 import { asArray as ColorAsArray } from 'ol/color';
 import { getVectorContext } from 'ol/render';
-import FormatType from 'ol/format/FormatType';
 import olFeature from 'ol/Feature';
 import olProjection from 'ol/proj/Projection';
 import * as olproj from 'ol/proj';
@@ -236,7 +235,7 @@ export class VectorLayer extends Layer {
   public centerMapOnFeature(id: string | number) {
     const feat = this.dataSource.ol.getFeatureById(id);
     if (feat) {
-      this.map.ol.getView().setCenter(feat.getGeometry().getCoordinates());
+      this.map.ol.getView().setCenter((feat.getGeometry() as any).getCoordinates());
     }
   }
 
@@ -425,9 +424,9 @@ export class VectorLayer extends Layer {
             const format = vectorSource.getFormat();
             const type = format.getType();
             let source;
-            if (type === FormatType.JSON || type === FormatType.TEXT) {
+            if (type === 'json' || type === 'text') {
               source = content;
-            } else if (type === FormatType.XML) {
+            } else if (type === 'xml') {
               source = content;
               if (!source) {
                 source = new DOMParser().parseFromString(
@@ -435,7 +434,7 @@ export class VectorLayer extends Layer {
                   'application/xml'
                 );
               }
-            } else if (type === FormatType.ARRAY_BUFFER) {
+            } else if (type === 'arraybuffer') {
               source = content;
             }
             if (source) {
@@ -451,7 +450,7 @@ export class VectorLayer extends Layer {
     } else {
     xhr.open( 'GET', modifiedUrl);
     const format = vectorSource.getFormat();
-    if (format.getType() === FormatType.ARRAY_BUFFER) {
+    if (format.getType() === 'arraybuffer') {
       xhr.responseType = 'arraybuffer';
     }
     if (interceptor) {
@@ -468,9 +467,9 @@ export class VectorLayer extends Layer {
       if (!xhr.status || (xhr.status >= 200 && xhr.status < 300)) {
         const type = format.getType();
         let source;
-        if (type === FormatType.JSON || type === FormatType.TEXT) {
+        if (type === 'json' || type === 'text') {
           source = xhr.responseText;
-        } else if (type === FormatType.XML) {
+        } else if (type === 'xml') {
           source = xhr.responseXML;
           if (!source) {
             source = new DOMParser().parseFromString(
@@ -478,7 +477,7 @@ export class VectorLayer extends Layer {
               'application/xml'
             );
           }
-        } else if (type === FormatType.ARRAY_BUFFER) {
+        } else if (type === 'arraybuffer') {
           source = xhr.response;
         }
         if (source) {
