@@ -7,7 +7,7 @@ import { map as rxMap } from 'rxjs/operators';
 import { saveAs } from 'file-saver';
 import jspdf from 'jspdf';
 import html2canvas from 'html2canvas';
-import * as JSZip from 'jszip';
+import { default as JSZip } from 'jszip';
 
 import { SubjectStatus } from '@igo2/utils';
 import { SecureImagePipe } from '@igo2/common';
@@ -19,6 +19,12 @@ import { LegendMapViewOptions } from '../../layer/shared/layers/layer.interface'
 import { getLayersLegends } from '../../layer/utils/outputLegend';
 
 import { PrintOptions } from './print.interface';
+
+declare global {
+  interface Navigator {
+    msSaveBlob?: (blob: any, defaultName?: string) => boolean
+  }
+}
 
 @Injectable({
   providedIn: 'root'
