@@ -15,7 +15,7 @@ import { DrawControl } from '../../geometry';
 import { createInteractionStyle, GeometryType } from '../../draw';
 import { featureToOl } from '../../feature';
 
-import type { default as OlGeometryType } from 'ol/geom/GeometryType';
+import type { Type } from 'ol/geom/Geometry';
 import type { default as OlGeometry } from 'ol/geom/Geometry';
 import OlGeoJSON from 'ol/format/GeoJSON';
 import OlVectorSource from 'ol/source/Vector';
@@ -227,7 +227,7 @@ export class EditionWorkspace extends Workspace {
    * Called when the user selects a new geometry type
    * @param geometryType the geometry type selected by the user
    */
-  onGeometryTypeChange(geometryType: typeof OlGeometryType, feature, workspace: EditionWorkspace) {
+  onGeometryTypeChange(geometryType: Type, feature, workspace: EditionWorkspace) {
       this.drawControl.setGeometryType(geometryType);
       this.toggleDrawControl(feature, workspace);
     }
